Fashion Item Classifier Predicton/app/requirements.txt
numpy==1.26.3
streamlit==1.30.0
tensorflow==2.15.1
Pillow==9.4.0
